Implement gh-3865
Replace unsigned long long types used to describe problem size with uint64_t
to avoid LP64/LLP64 ambiguity.

+++ c/parallel/src/reduce.cu
@@ -387,7 +387,7 @@ CUresult cccl_device_reduce(
   size_t* temp_storage_bytes,
   cccl_iterator_t d_in,
   cccl_iterator_t d_out,
-  unsigned long long num_items,
+  uint64_t num_items,
   cccl_op_t op,
   cccl_value_t init,
   CUstream stream)
